requirements.txt
django<=1.6.6
django-guardian
boto
django-social-auth
django-bootstrap3
djangorestframework<=2.4
drf-compound-fields
markdown
django-filter
south
django-debug-toolbar
django-encrypted-fields
MySQL-python
django-reversion<=1.8
